pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "calliope"
version = "0.1.0"
description = "Git-based version control for fiction writers with AI assistance."
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Your Name"}]
dependencies = [
    "gitpython>=3.1.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "python-docx>=1.0.0",
    "markdown>=3.4.0",
]

[project.scripts]
calliope = "calliope.cli:main"

[tool.setuptools]
include-package-data = true
